Sudoku solver
The main goal of this project is to become more familiar with C++, data structures, and algorithms. All the sudoku code is represented in sudoku.txt file. A single comment sign '/' is line ignoring trigger. If there is 1 error in sudoku filling, the algorithm will replace it and solve it, otherwise - error will be shown (that sudoku data must be edited).